+++ .github/workflows/release.yml
@@ -2,7 +2,7 @@ name: Release
 
 on:
   push:
-    branches: [ "Develop" ]
+    branches: [ "develop" ]
 
 
 #these job is for build (remember we are creating another job for test)
